Nominal Parameter Calculation Addition
Added a nominal parameter calculation which is used during model optimization. Parameters that are not optimized are set at patient specific nominal values. Also changed ao and vc subscripts to sa and sv to reflect that the compartments are systemic arteries and veins not aorta an vena cava. Also added additional comments to the code