Single Simulation Run Model Code
All optimization options are stripped out of this code along with plotting options comparing simulation to normal CV function on the same graph. This leaves a model code that runs a simulation for patient 27 in the manuscript and allows the user to change parameters in the script to hand tune the model to the users own input data.